src/finetuning/llm-finetuning/gemini-finetuner/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.0",
    "google-cloud-storage>=3.4.1",
    "google-genai>=1.39.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "openai>=1.95.1",
    "openai-agents>=0.3.2",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
]
